♻️ refactor: Change endpoints
