+++ division-service/src/main/java/com/bytogether/divisionservice/controller/DivisionController.java
@@ -23,36 +23,4 @@ public class DivisionController {
         return divisionService.getDivisionByCoordinates(coordinate);
     }
 
-    // 읍면동 리스트 검색 (좌표로) - 완성 (최적화는 이후에 현재 쿼리 여러번 날림 - 250923 14:26)
-    @PostMapping("/private/list/by-coord")
-    public List<Optional<Division>> getDivisionListByCoordinatesList(@RequestBody @Valid Coordinates coordinates) {
-        return divisionService.getDivisionListByCoordinatesList(coordinates);
-    }
-
-
-    // 읍면동 검색 (읍면동 코드로) - 완성
-    @GetMapping("/private/by-code")
-    public Optional<Division> getDivisionByCode(@Valid Emd emd) {
-        return divisionService.getDivisionByCode(emd);
-    }
-
-    //읍면동 리스트 검색 (읍면동 코드로) - 완성 (최적화는 이후에 현재 쿼리 여러번 날림 - 250923 14:26)
-    @PostMapping("/private/list/by-code")
-    public List<Optional<Division>> getDivisionListByCodeList(@RequestBody @Valid Emds emds) {
-        return divisionService.getDivisionListByCodeList(emds);
-    }
-
-
-    // 인접동 검색 (읍면동 코드로) - 완성
-    @GetMapping("/private/near/by-code")
-    public List<Division> getNearDivisionListByCode(@Valid EmdDepth emdDepth) {
-        return divisionService.getNearDivisionByCode(emdDepth);
-    }
-
-    // 인접동 검색 (좌표로) - 완성
-    @GetMapping("/private/near/by-coord")
-    public List<Division> getNearDivisionListByCoordinates(@Valid CoordinateDepth coordinateDepth) {
-        return divisionService.getNearDivisionByCoordinates(coordinateDepth);
-    }
-
 }
